requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
openai
python-dotenv
python-multipart